Split runtime live sidecar tests
